fox/requirements.txt
asgiref==3.2.10
Django==3.0.4
django-crispy-forms==1.9.2
Pillow==7.2.0
psycopg2-binary==2.8.5
python-slugify==4.0.1
pytz==2020.1
sqlparse==0.3.1
text-unidecode==1.3
